+++ server/modules/snark-worker.js
@@ -12,6 +12,7 @@ const processSnarkWorkerController = async () => {
     const host = hostname()
 
     if (!swConfig) return
+    if (globalThis.snarkWorkerRunError) return
 
     const {
         address,
@@ -37,6 +38,9 @@ const processSnarkWorkerController = async () => {
         cmdFee = setWorkerFeeCommand.replace("<FEE>", fee)
         setFee = execCommand(cmdFee)
         setFee.on("exit", (code) => {
+            if (code === 0) {
+                logging(`Snark Worker Fee ${fee} successfully installed!`)
+            }
             sendAlert("EXEC", `Command \`${cmdFee}\` executed ${code === 0 ? "successfully" : "with error code " + code} for the host ${host}`)
         })
 
@@ -46,6 +50,10 @@ const processSnarkWorkerController = async () => {
             if (code === 0) {
                 globalThis.snarkWorkerStopped = false
                 globalThis.snarkWorkerStoppedBlockTime = null
+                logging("Snark Worker successfully started!")
+            } else {
+                globalThis.snarkWorkerRunError = true
+                sendAlert("EXEC", "Snark Work run error! Snark Worker Controller was stopped!")
             }
             sendAlert("EXEC", `Command \`${cmdStart}\` executed ${code === 0 ? "successfully" : "with error code " + code} for the host ${host}`)
         })
@@ -59,6 +67,7 @@ const processSnarkWorkerController = async () => {
             if (code === 0) {
                 globalThis.snarkWorkerStopped = true
                 globalThis.snarkWorkerStoppedBlockTime = nextBlock
+                logging("Snark Worker successfully stopped!")
             }
             sendAlert("EXEC", `Command ${cmdStop} executed ${code === 0 ? "successfully" : "with error code " + code} for the host ${host}`)
         })