+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 2.0.1
 + [x] Server: fixed snark worker controller for first start, issue #7
++ [x] Server: fixed spams when snark worker controller can not start sw
 
 ### 2.0.0
 + [x] Server: changed data exchange protocol from `http(s)` to `ws(s)`